pyproject.toml
[project]
name = "crag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "black>=25.1.0",
    "isort>=6.0.0",
    "langchain>=0.3.19",
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.3.18",
    "langchain-openai>=0.3.9",
    "langchainhub>=0.1.21",
    "langgraph>=0.2.74",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.5.1",
]
